+++ chats/apps/api/v1/dashboard/repository.py
@@ -21,6 +21,7 @@ from .dto import (
     QueueRoomData,
     Sector,
     TransferRoomData,
+    get_admin_domains_exclude_filter,
 )
 from .interfaces import CacheRepository, RoomsDataRepository
 
@@ -67,7 +68,7 @@ class AgentRepository:
 
         agents_query = self.model
         if not filters.is_weni_admin:
-            agents_query = agents_query.exclude(email__endswith="weni.ai")
+            agents_query = agents_query.exclude(get_admin_domains_exclude_filter())
 
         agents_query = (
             agents_query.filter(project_permissions__project=project, is_active=True)