+++ chats/apps/rooms/views.py
@@ -54,21 +54,6 @@ def get_editable_custom_fields_room(room_filter: dict) -> Room:
     return room
 
 
-def create_transfer_json(action: str, from_, to):
-    if hasattr(from_, "_meta"):
-        return {
-            "action": action,
-            "from": {"type": from_._meta.model_name, "name": from_.name},
-            "to": {"type": to._meta.model_name, "name": to.name},
-        }
-    else:
-        return {
-            "action": action,
-            "from": {"type": "", "name": ""},
-            "to": {"type": to._meta.model_name, "name": to.name},
-        }
-
-
 def create_feedback_json(method: str, content: dict):
     return {"method": method, "content": content}